refactor: domain, view, service, main(controller) 로 리패키징
