+++ src/main/kotlin/lotto/vo/WinningLottoPrizeVO.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.vo
 
 import java.math.BigDecimal
