+++ src/test/kotlin/lotto/domain/LottosTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
