+++ src/test/kotlin/lotto/domain/PrizeTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.datatest.withData
